port binja integration to aglib
