+++ pwndbg/integration/binja.py
@@ -20,7 +20,6 @@ from typing import List
 from typing import Tuple
 from typing import TypeVar
 
-import gdb
 import pygments
 import pygments.formatters
 import pygments.style
@@ -43,7 +42,9 @@ from pwndbg.aglib.nearpc import c as nearpc_color
 from pwndbg.aglib.nearpc import ljust_padding
 from pwndbg.color import message
 from pwndbg.color import theme
+from pwndbg.dbg import BreakpointLocation
 from pwndbg.dbg import EventType
+from pwndbg.dbg import StopPoint
 from pwndbg.lib.functions import Argument
 from pwndbg.lib.functions import Function
 
@@ -219,7 +220,7 @@ def auto_update_pc() -> None:
     _bn.update_pc_tag(l2r(pc))
 
 
-_managed_bps: Dict[int, gdb.Breakpoint] = {}
+_managed_bps: Dict[int, StopPoint] = {}
 
 
 @pwndbg.dbg.event_handler(EventType.START)
@@ -232,9 +233,12 @@ def auto_update_bp() -> None:
     bps: List[int] = _bn.get_bp_tags()
     binja_bps = {r2l(addr) for addr in bps}
     for k in _managed_bps.keys() - binja_bps:
-        _managed_bps.pop(k).delete()
+        bp = _managed_bps.pop(k)
+        bp.remove()
+
+    inf = pwndbg.dbg.selected_inferior()
     for k in binja_bps - _managed_bps.keys():
-        bp = gdb.Breakpoint("*" + hex(k))
+        bp = inf.break_at(BreakpointLocation(k))
         _managed_bps[k] = bp
 
 
@@ -527,7 +531,7 @@ class BinjaProvider(pwndbg.integration.IntegrationProvider):
         newest = True
         # try to find the oldest frame that's earlier than the address
         while True:
-            upper = gdb_frame_to_dbg(dbg_frame_to_gdb(cur).older())
+            upper = cur.parent()
             if upper is None:
                 break
 
@@ -556,21 +560,3 @@ class BinjaProvider(pwndbg.integration.IntegrationProvider):
             return f"{var}{suffix}"
         else:
             return f"{func}:{var}{suffix}"
-
-
-def dbg_frame_to_gdb(d: pwndbg.dbg_mod.Frame) -> gdb.Frame:
-    # TODO: fix later to aglib
-    from pwndbg.dbg.gdb import GDBFrame
-
-    assert isinstance(d, GDBFrame)
-    return d.inner
-
-
-def gdb_frame_to_dbg(d: gdb.Frame | None) -> pwndbg.dbg_mod.Frame | None:
-    # TODO: fix later to aglib
-    from pwndbg.dbg.gdb import GDBFrame
-
-    if d is None:
-        return None
-
-    return GDBFrame(d)
